Refactor into npm workspaces
