+++ package.json
@@ -3,21 +3,11 @@
   "version": "0.2.0",
   "private": true,
   "scripts": {
-    "start": "npm run build-data && react-scripts start",
-    "build": "npm run build-data && react-scripts build",
-    "build-data": "node ./src/compileData.js"
+    "start": "npm run dev --workspace=frontend",
+    "build": "npm run build-data && npm run build --workspace=frontend",
+    "build-data": "node .packages/scripts/compileData.js"
   },
   "dependencies": {
-    "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.3.2",
-    "@testing-library/user-event": "^14.6.1",
-    "embed-js": "^5.0.4",
-    "lodash": "^4.17.23",
-    "prop-types": "^15.8.1",
-    "pym.js": "^1.3.2",
-    "react": "^19.2.4",
-    "react-dom": "^19.2.4",
-    "styled-components": "^6.3.9",
     "xlsx": "^0.18.5"
   },
   "devDependencies": {
@@ -32,6 +22,7 @@
     "typescript": "^5.9.3",
     "typescript-eslint": "^8.54.0"
   },
+  "workspaces": ["packages/*"],
   "browserslist": {
     "production": [
       ">0.2%",
