+++ packages/frontend/public/embed.js
