+++ packages/frontend/public/favicons/android-chrome-144x144.png
