+++ packages/frontend/public/favicons/android-chrome-192x192.png
