+++ packages/frontend/public/favicons/android-chrome-36x36.png
