+++ packages/frontend/public/favicons/android-chrome-72x72.png
