+++ packages/frontend/public/favicons/android-chrome-96x96.png
