+++ packages/frontend/public/favicons/apple-touch-icon-114x114.png
