+++ packages/frontend/public/favicons/apple-touch-icon-120x120.png
