+++ packages/frontend/public/favicons/apple-touch-icon-144x144.png
