+++ packages/frontend/public/favicons/apple-touch-icon-152x152.png
