+++ packages/frontend/public/favicons/apple-touch-icon-180x180.png
