+++ packages/frontend/public/favicons/apple-touch-icon-57x57.png
