+++ packages/frontend/public/favicons/apple-touch-icon-60x60.png
