+++ packages/frontend/public/favicons/apple-touch-icon-72x72.png
