+++ packages/frontend/public/favicons/apple-touch-icon-76x76.png
