+++ packages/frontend/public/favicons/apple-touch-icon-precomposed.png
