+++ packages/frontend/public/favicons/apple-touch-icon.png
