+++ packages/frontend/public/favicons/browserconfig.xml
