+++ packages/frontend/public/favicons/favicon-32x32.png
