+++ packages/frontend/public/favicons/favicon-96x96.png
