+++ packages/frontend/public/favicons/favicon.ico
