+++ packages/frontend/public/favicons/mstile-310x310.png
