+++ packages/frontend/public/favicons/mstile-70x70.png
