+++ packages/frontend/public/logo192.png
