+++ packages/frontend/public/logo512.png
