+++ packages/frontend/public/manifest.json
