+++ packages/scripts/parseStateTable.js
@@ -1,6 +1,6 @@
 const _ = require('lodash');
 
-const states = require('./states.json');
+const states = require('../frontend/src/data/states.json');
 
 module.exports = function(table) {
   /*
